// generated/mapsquare_locations.d.ts
// GENERATED DO NOT EDIT
// This source data is located at '..\src\opcodes\mapsquare_locations.json'
// run `npm run filetypes` to rebuild

export type mapsquare_locations = {
	locations: {
		id: number,
		uses: {
			y: number,
			x: number,
			plane: number,
			rotation: number,
			type: number,
			extra: {
				flags: number,
				rotation: number[] | undefined,
				translateX: number | undefined,
				translateY: number | undefined,
				translateZ: number | undefined,
				scale: number | undefined,
				scaleX: number | undefined,
				scaleY: number | undefined,
				scaleZ: number | undefined,
			} | undefined,
		}[],
	}[],
};